+++ ethcore/sync/src/chain/handler.rs
@@ -33,7 +33,7 @@ use types::BlockNumber;
 use types::block_status::BlockStatus;
 use types::ids::BlockId;
 
-use super::packet::{Packet, SyncPacketId, SyncPacketId::*};
+use super::syncpacketid::{Packet, SyncPacketId, SyncPacketId::*};
 use super::{
 	BlockSet,
 	ChainSync,
