+++ ethcore/sync/src/chain/mod.rs
@@ -88,7 +88,7 @@
 //! All other messages are ignored.
 
 mod handler;
-pub mod packet;
+pub mod syncpacketid;
 mod propagator;
 mod requester;
 mod supplier;
@@ -119,7 +119,7 @@ use types::transaction::UnverifiedTransaction;
 use types::BlockNumber;
 
 use self::handler::SyncHandler;
-use self::packet::{Packet, SyncPacketId, SyncPacketId::*};
+use self::syncpacketid::{Packet, SyncPacketId, SyncPacketId::*};
 use self::propagator::SyncPropagator;
 use self::requester::SyncRequester;
 pub(crate) use self::supplier::SyncSupplier;
