+++ ethcore/sync/src/chain/syncpacketid.rs
